backend/requirements.txt
fastapi
uvicorn
opencv-python
numpy
pydantic
python-dotenv==1.0.1